More work on updating libraries

+++ ccam/main.lua
@@ -32,9 +32,10 @@ function downloadApp(app_name)
 	f_sc.close()
 end
 
-function deleteApp(app_name)
+function deleteApp(app_name, isLib)
+	local dir = isLib and CCAM_CONF.LIB_DIR or CCAM_CONF.APP_DIR
 	-- Check that app exists
-	if not appExists(app_name) then
+	if not appExists(app_name, isLib) then
 		printError("Error: App doesn't exist")
 		return false
 	end
@@ -44,8 +45,10 @@ function deleteApp(app_name)
 	local ans = read()
 	if ans == 'y' or ans == 'Y' then
 		-- Remove the app
-		fs.delete(CCAM_CONF.APP_DIR .. app_name)
-		fs.delete(CCAM_CONF.BIN_DIR .. app_name)
+		fs.delete(dir .. app_name)
+		if isLib then
+			fs.delete(CCAM_CONF.BIN_DIR .. app_name)
+		end
 	else
 		print("Aborted.")
 	end
@@ -57,7 +60,7 @@ function updateApp(app_name, isLib)
 	local dir = isLib and CCAM_CONF.LIB_DIR or CCAM_CONF.APP_DIR
 
 	-- Check that app exists
-	if not appExists(app_name) then
+	if not appExists(app_name, isLib) then
 		printError("Error: App doesn't exist")
 		return false
 	end
@@ -143,6 +146,7 @@ function getAppVersion(app_name, isLib)
 	return data.version
 end
 
-function appExists(app_name)
-	return fs.exists(CCAM_CONF.APP_DIR .. app_name .. CCAM_CONF.APP_MAIN)
+function appExists(app_name, isLib)
+	local dir = isLib and CCAM_CONF.LIB_DIR or CCAM_CONF.APP_DIR
+	return fs.exists(dir .. app_name)
 end
